+ улучшение изображения

+++ FaceDetection/FaceDetection/Human.cs
@@ -11,13 +11,13 @@ namespace FaceDetection
     {
         public int Id { get; private set; }
         public string Name { get; private set; }
-        public List<Image<Bgr, float>> Images { get; private set; }
+        public List<Image<Bgr, byte>> Images { get; private set; }
 
         public Human(int id, string name)
         {
             this.Id = id;
             this.Name = name;
-            this.Images = new List<Image<Bgr, float>>();
+            this.Images = new List<Image<Bgr, byte>>();
         }
     }
 }
